Create speed_pass.cpp
Speed Pass provides 2-pass security. It will protect your data even if your password is known to someone. It firstly matches the string entered with the original password.
If it get a pass through then it will check for the typing speed of the user. If this speed matches the typing speed of the authorized user, it will provide him the instant access.